backend/requirements.txt
asgiref==3.4.1
dj-database-url==0.5.0
Django==3.2.5
django-cors-headers==3.7.0
django-environ==0.4.5
django-filter==2.4.0
django-heroku==0.3.1
djangorestframework==3.12.4
djongo==1.3.6
dnspython==2.1.0
gunicorn==20.1.0
importlib-metadata==4.6.1
Markdown==3.3.4
pymongo==3.11.4
pytz==2021.1
sqlparse==0.2.4
typing-extensions==3.10.0.0
whitenoise==5.2.0
zipp==3.5.0
